Collecting all hardcoded constants in constants.json
Constants.json also contains defaults for the options that can be
configured in the config file.
The backend & frontend read constants.json and extract the constants
they need in constants.py and constants.js respectively.

Added replicated_cameras as an extra config option.
As in the previous commit, core.py parses the config file. Any
constants defined in the config file replace the constants defined in
constants.json.

Also fix on constants import in common.js