Adding tab that display instGraph diagram.
The full path to the instGraph file is set with the config variable instgraph_file_path.
On the backend, starting 'sup' also starts a watchdog thread that monitors the file that instgraph_file_path points to.
If a change is detected, a message with the filename is sent via websocket to the frontend.
The watchdog handler also implements a basic debouncer such that no message is sent via websocket for updates more frequent than the given
debouncing_interval_sec (by default 0.2s, but modifiable via the config file).
There is also a new mounted '/drawio-files' path, which is an efficient way for the frontend to request the file.

The frontend has two new Vuex.store entries: instGraphUpdateTime and instGraphFilename.
When the websocket message is received they are both updated.
The new instGraph component watches for changes on both of these and when a change is detected it makes a request to
/drawio-files/instGraphFilename, parses the xml and uses maxgraph to display it.